concierge-agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
openai==1.46.0
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.39.0
loguru==0.7.2
pydantic==2.9.2
-e .